added support for kubernetes endpoint (#6697)
* added support for kubernetes endpoint

* Update task.loc.json
